third_party/WebKit/LayoutTests/http/tests/websocket/tests/hybi/connect-error-by-no-websocket-server-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8888/' failed: WebSocket is closed before the connection is established.
Test that WebSocket fires error event if no WebSocket Server can be connected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test0 Start
PASS onerror was called
PASS onclose was called
test1 Start
PASS onerror was called
PASS onclose was called
test2 Start
PASS onerror was called
PASS onclose was called
PASS successfullyParsed is true

TEST COMPLETE
